Add PS1 "20 queries" filtering notebook
Add a notebook containing a subset of "20 queries"  for PanSTARRS, adapted from the queries for SDSS presented by Gray et al. 2002. This notebook covers 3 queries (Q2, Q4, Q5) which can be accomplished through table joins and filtering.

Also added a "aggregate" overview page for the PS1 "20 query" tutorial, following the style of the SDSS notebooks.

This "PS1 landing page" could be modified to include the other two PS1 notebooks, if that organization structure would be preferred.